MAINT: moved much from `Kernel` back to `Likelihood`
