+++ mosfit/modules/kernels/kernel.py
@@ -1,13 +1,11 @@
 """Definitions for the `Kernel` class."""
 from collections import OrderedDict
-from math import isnan
 
 import numpy as np
 from six import string_types
 
 from mosfit.constants import ANG_CGS, C_CGS
 from mosfit.modules.module import Module
-from mosfit.utils import flux_density_unit
 
 
 # Important: Only define one ``Module`` class per file.
@@ -18,18 +16,16 @@ class Kernel(Module):
 
     MIN_COV_TERM = 1.0e-30
 
+    def __init__(self, **kwargs):
+        """Initialize module."""
+        super(Kernel, self).__init__(**kwargs)
+        self._times = np.array([])
+        self._full = kwargs.get('full', False)
+
     def process(self, **kwargs):
         """Process module."""
         self.preprocess(**kwargs)
-        self._model_observations = kwargs['model_observations']
-        self._fractions = kwargs['fractions']
         ret = {}
-        if min(self._fractions) < 0.0 or max(self._fractions) > 1.0:
-            return ret
-        for oi, obs in enumerate(self._model_observations):
-            if not self._upper_limits[oi] and (isnan(obs) or
-                                               not np.isfinite(obs)):
-                return ret
         self._codeltatime = kwargs.get(self.key('codeltatime'), -1)
         self._codeltalambda = kwargs.get(self.key('codeltalambda'), -1)
 
@@ -58,68 +54,18 @@ class Kernel(Module):
 
         self._o_band_vs = self._band_vs[self._observed]
 
-        self._model_observations[self._cmask] = -2.5 * np.log10(
-            self._model_observations[self._cmask])
-
-        # Calculate (model - obs) residuals.
-        residuals = np.array([
-            (abs(x - ct) if not u or (x < ct and not isnan(x)) else 0.0)
-            if t == 'countrate' and ct is not None
-            else
-            (abs(x - y) if not u or (x < y and not isnan(x)) else 0.0)
-            if t == 'magnitude' and y is not None
-            else
-            (abs(x - fd) if not u or (x > fd and not isnan(x)) else 0.0)
-            if t == 'fluxdensity' and fd is not None else None
-            for x, y, ct, fd, u, o, t in zip(
-                self._model_observations, self._mags, self._cts, self._fds,
-                self._upper_limits, self._observed, self._observation_types)
-            if o
-        ])
-
-        if np.any(residuals == None):  # noqa: E711
-            raise ValueError('Null residual.')
-
-        # Observational errors to be put in diagonal of error matrix.
-        diag = np.array([
-            ((ctel if x > ct else cteu) ** 2)
-            if t == 'countrate' and ct is not None else
-            ((el if x > y else eu) ** 2)
-            if t == 'magnitude' and y is not None else
-            ((fdel if x < fd else fdeu) ** 2)
-            if t == 'fluxdensity' and fd is not None else None
-            for x, y, eu, el, fd, fdeu, fdel, ct, ctel, cteu, o, t in zip(
-                self._model_observations, self._mags,
-                self._e_u_mags, self._e_l_mags, self._fds, self._e_u_fds,
-                self._e_l_fds, self._cts, self._e_l_cts, self._e_u_cts,
-                self._observed, self._observation_types) if o
-        ])
-
-        if np.any(diag == None):  # noqa: E711
-            raise ValueError('Null error.')
-
-        is_diag = False
         if self._codeltatime >= 0 or self._codeltalambda >= 0:
             kmat = np.outer(self._o_band_vs, self._o_band_vs)
-        else:
-            # Shortcut when matrix is diagonal.
-            is_diag = True
-            value = -0.5 * np.sum(
-                residuals ** 2 / (self._o_band_vs ** 2 + diag) +
-                np.log(self._o_band_vs ** 2 + diag))
 
-            ret['value'] = value
-
-        if not is_diag:
             if self._codeltatime >= 0:
                 kmat *= np.exp(self._dt2mat / self._codeltatime ** 2)
 
             if self._codeltalambda >= 0:
                 kmat *= np.exp(self._dl2mat / self._codeltalambda ** 2)
 
-            ret['diag'] = diag
             ret['kmat'] = kmat
-            ret['residuals'] = residuals
+        else:
+            ret['obandvs'] = self._o_band_vs
 
         return ret
 
@@ -129,103 +75,22 @@ class Kernel(Module):
         self._variance_bands = requests.get('variance_bands', [])
 
     def preprocess(self, **kwargs):
-        """Construct arrays of observations based on data keys."""
-        if self._preprocessed:
+        """Construct kernel distance arrays."""
+        new_times = np.array(kwargs.get('all_times', []), dtype=float)
+        if np.array_equiv(new_times, self._times) and self._preprocessed:
             return
-        self._times = np.array(kwargs.get('times', []))
-        self._mags = kwargs.get('magnitudes', [])
-        self._fds = kwargs.get('fluxdensities', [])
-        self._cts = np.array(kwargs.get('countrates', []))
-        self._freqs = kwargs.get('frequencies', [])
-        self._e_u_mags = kwargs.get('e_upper_magnitudes', [])
-        self._e_l_mags = kwargs.get('e_lower_magnitudes', [])
-        self._e_mags = kwargs.get('e_magnitudes', [])
-        self._e_u_fds = kwargs.get('e_upper_fluxdensities', [])
-        self._e_l_fds = kwargs.get('e_lower_fluxdensities', [])
-        self._e_fds = kwargs.get('e_fluxdensities', [])
-        self._u_fds = kwargs.get('u_fluxdensities', [])
-        self._u_freqs = kwargs.get('u_frequencies', [])
-        self._e_u_cts = kwargs.get('e_upper_countrates', [])
-        self._e_l_cts = kwargs.get('e_lower_countrates', [])
-        self._e_cts = kwargs.get('e_countrates', [])
-        self._u_cts = kwargs.get('u_countrates', [])
-        self._upper_limits = kwargs.get('upperlimits', [])
-        self._observed = np.array(kwargs.get('observed', []))
+        self._times = new_times
         self._all_band_indices = kwargs.get('all_band_indices', [])
-        self._observation_types = kwargs['observation_types']
         self._are_bands = np.array(self._all_band_indices) >= 0
         self._all_band_avgs = np.array([
             self._average_wavelengths[bi] if bi >= 0 else
             C_CGS / self._freqs[i] / ANG_CGS for i, bi in
             enumerate(self._all_band_indices)])
+        if self._full:
+            self._observed = np.full(len(self._times), True)
+        else:
+            self._observed = np.array(kwargs.get('observed', []), dtype=bool)
         self._n_obs = len(self._observed)
-
-        # Magnitudes first
-        # Note: Upper limits (censored data) currently treated as a
-        # half-Gaussian, this is very approximate and can be improved upon.
-        self._e_u_mags = [
-            kwargs['default_upper_limit_error']
-            if (e is None and eu is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and eu is None) else (e if eu is None else eu))
-            for i, (e, eu) in enumerate(zip(self._e_mags, self._e_u_mags))
-        ]
-        self._e_l_mags = [
-            kwargs['default_upper_limit_error']
-            if (e is None and el is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and el is None) else (e if el is None else el))
-            for i, (e, el) in enumerate(zip(self._e_mags, self._e_l_mags))
-        ]
-
-        # Now counts
-        self._cmask = self._observation_types == 'countrate'
-        self._cts[self._cmask] = -2.5 * np.log10(self._cts[self._cmask]
-                                                 .astype(np.float64))
-        self._e_u_cts = [
-            kwargs['default_upper_limit_error']
-            if (e is None and eu is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and eu is None) else
-             2.5 * (np.log10(c + (e if eu is None else eu)) - np.log10(c)))
-            for i, (c, e, eu) in enumerate(zip(
-                self._cts, self._e_cts, self._e_u_cts))
-        ]
-        self._e_l_cts = [
-            kwargs['default_upper_limit_error']
-            if (e is None and el is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and el is None) else
-             2.5 * (np.log10(c) - np.log10(c - (e if el is None else el))))
-            for i, (c, e, el) in enumerate(zip(
-                self._cts, self._e_cts, self._e_l_cts))
-        ]
-
-        # Now flux densities
-        self._e_u_fds = [
-            v if (e is None and eu is None and self._upper_limits[i]) else
-            (v if (e is None and eu is None) else (e if eu is None else eu))
-            for i, (e, eu, v) in enumerate(
-                zip(self._e_fds, self._e_u_fds, self._fds))
-        ]
-        self._e_l_fds = [
-            0.0 if self._upper_limits[i] else (
-                v if (e is None and el is None) else (e if el is None else el))
-            for i, (e, el, v) in enumerate(
-                zip(self._e_fds, self._e_l_fds, self._fds))
-        ]
-        self._fds = [
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._fds, self._u_fds)
-        ]
-        self._e_u_fds = [
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._e_u_fds, self._u_fds)
-        ]
-        self._e_l_fds = [
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._e_l_fds, self._u_fds)
-        ]
 
         self._o_times = self._times[self._observed]
         self._o_waves = self._all_band_avgs[self._observed]
